chore(api, types): move `getExpressionContext` and `expressionContextSchema` to dictionary module
